Enhance Atmel SPI implementation
* Allow dummy character to be set; this character gets transmitted when
there is no configured transmit buffer.

Note: The SAMD21 slave SPI has some issues which this feature
exacerbates.  The shift register gets preloaded when SS is driven low,
allowing for the first transmitted character to be correct.  The problem
is that there is no way to clear the shift register.  The shift register
only gets overwritten when the master sends another byte.  Consequently,
slave mode is really only useful for this MCU when you know exactly how
many characters the master will be transmitting.  Otherwise, if you try
to compensate for a large command by transmitting a too-large buffer,
the first unsent character from the buffer will get sent upon receiving
the second command.

This problem is exacerbated by the dummy character feature.  If you set
a dummy character and then attempt a transmit, the first transmitted
character will be the dummy character that was preloaded into the shift
register!